config: move the CoS code-point resolver to its own file
The #7084 addition pushed compiler_class_of_service.go from 1462 to 1520
LOC, crossing the 1500 [WATCH] floor, and
TestTouchedFileCrossedModularityThreshold red on it. That gate offers
two outs -- split the file, or record the decision in
docs/refactoring-audit-accepted.txt -- and splitting is the right one
here: the added code is a token->value resolver plus its alias table,
with no dependency on anything else in the file.

The new file is named for a seam this package already recognises:
cos_code_points_spellings_6697_test.go is its test counterpart. The
DSCP alias table and the two collectors still in
compiler_class_of_service.go belong here as well, and the comment says
so, but moving them belongs in a change whose SUBJECT is that motion --
doing it here would bury four unrelated CoS defects inside a refactor.

compiler_class_of_service.go returns to 1463 LOC, below the floor, and
pkg/refactoraudit is green.
